backend/requirements.txt
# FastAPI и ASGI сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
authlib==1.3.0
itsdangerous==2.1.2

# Email
fastapi-mail==1.4.1
jinja2==3.1.3

# Telegram
python-telegram-bot==20.7
aiogram==3.3.0
# Note: python-telegram-bot requires httpx, version managed by it

# Парсинг
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# Валидация и утилиты
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
pillow==10.2.0

# Тестирование
pytest==7.4.4
pytest-asyncio==0.23.3

# Code quality
ruff==0.1.11
mypy==1.8.0

# Прочее
python-slugify==8.0.1
